Expose getStorefront APIs in CustomEntitlementComputation flavor
